Fix issues

+++ server_litestar/app/controllers/items.py
@@ -1,16 +1,18 @@
 """Items routes."""
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
-from litestar import Controller, delete, get, patch, post
+from litestar import Controller, Request, delete, get, patch, post
+from litestar.exceptions import NotFoundException
+from litestar.security.jwt import Token
 from pydantic import TypeAdapter
 from sqlalchemy import label, literal, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.classes.classes import ItemCategory, Roles
-from app.middleware.auth import role_guard
+from app.middleware.auth import User, role_guard
 from app.models.items import ItemModelIn, ItemsOutModels, ItemTypeOut
-from app.tables.tables import tables
+from app.tables.tables import Persons, tables
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -25,7 +27,7 @@ class ItemsController(Controller):
 
     @staticmethod
     async def select_item(
-        item: ItemCategory,
+        item: str,
         person_id: int,
         db_session: AsyncSession,
     ) -> Sequence:
@@ -47,6 +49,15 @@ class ItemsController(Controller):
             .order_by(table.c.id.desc())
         )
         return (await db_session.execute(stmt)).all()
+
+    @staticmethod
+    async def get_user_id(
+        person_id: int,
+        db_session: AsyncSession,
+    ) -> int:
+        """Get user ID assotiated with person."""
+        stmt = select(Persons.user_id).filter(Persons.id == person_id)
+        return (await db_session.execute(stmt)).scalar_one()
 
     @get("/{person_id:int}")
     async def get_items(
@@ -100,6 +111,7 @@ class ItemsController(Controller):
         self,
         person_id: int,
         data: ItemModelIn,
+        request: Request[User, Token, Any],
         db_session: AsyncSession,
     ) -> None:
         """Add a new item to the database.
@@ -107,16 +119,21 @@ class ItemsController(Controller):
         Args:
             person_id: Person ID.
             data: ItemModelIn.
+            request: Request.
             db_session: AsyncSession.
 
         Returns:
             Response with status code 201.
 
         """
-        item = data.item.item
-        json_dict = data.item.model_dump(exclude={"item"}) | {"person_id": person_id}
-        stmt = tables[item].insert().values(json_dict)
-        await db_session.execute(stmt)
+        if self.get_user_id(person_id, db_session) == request.user.id:
+            item = data.item.item
+            json_dict = data.item.model_dump(exclude={"item"}) | {
+                "person_id": person_id,
+            }
+            stmt = tables[item].insert().values(json_dict)
+            await db_session.execute(stmt)
+        raise NotFoundException
 
     @patch(
         "/{person_id:int}/{item_id:int}",
@@ -128,6 +145,7 @@ class ItemsController(Controller):
         person_id: int,
         item_id: int,
         data: ItemModelIn,
+        request: Request[User, Token, Any],
         db_session: AsyncSession,
     ) -> None:
         """Update an item in the database.
@@ -136,38 +154,47 @@ class ItemsController(Controller):
             person_id: Person ID.
             item_id: Item ID.
             data: ItemModelIn.
+            request: Request.
             db_session: AsyncSession.
 
         Returns:
             Response with status code 201.
 
         """
-        json_dict = data.item.model_dump() | {"person_id": person_id}
-        table = tables[json_dict.pop("item")]
-        stmt = table.update().where(table.c.id == item_id).values(json_dict)
-        await db_session.execute(stmt)
+        if self.get_user_id(person_id, db_session) == request.user.id:
+            json_dict = data.item.model_dump() | {"person_id": person_id}
+            table = tables[json_dict.pop("item")]
+            stmt = table.update().where(table.c.id == item_id).values(json_dict)
+            await db_session.execute(stmt)
+        raise NotFoundException
 
     @delete(
-        "/{item:str}/{item_id:int}",
+        "/{item:str}/{person_id:int}/{item_id:int}",
         guards=[role_guard],
         opt={"role": Roles.user.value},
     )
     async def delete_item(
         self,
         item: ItemCategory,
+        person_id: int,
         item_id: int,
+        request: Request[User, Token, Any],
         db_session: AsyncSession,
     ) -> None:
         """Delete an item from the database.
 
         Args:
             item: ItemCategory.
+            person_id: Person ID.
             item_id: Item ID.
+            request: Request.
             db_session: AsyncSession.
 
         Returns:
             Response with status code 204.
 
         """
-        table = tables[item]
-        await db_session.execute(table.delete().where(table.c.id == item_id))
+        if self.get_user_id(person_id, db_session) == request.user.id:
+            table = tables[item]
+            await db_session.execute(table.delete().where(table.c.id == item_id))
+        raise NotFoundException
